+++ engine/source/engine/text_size.cpp
@@ -12,7 +12,7 @@ void update_text_sizes()
 	// It's a linear function.
 	// 768  =>  12
 	// 1080 =>  16
-	Context.text_sizes[Text_Size::Developer_Console] =( 1/72.0f ) * H + ( 28 / 13.0f );
+	Context.text_sizes[Text_Size::Developer_Console] = static_cast<s8>( ( 1/72.0f ) * H + ( 28 / 13.0f ) );
 	Context.text_sizes[Text_Size::Menu_Button] = 72;
 }
 }
